refactor: remove unused types for kiwi and user, and relocate relevant interfaces to idb utility
