--- apps/web/src/types/user.ts
@@ -1,6 +1,0 @@
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  profileImage: string;
-}
